requirements.txt
scrapy>=2.13.0
beautifulsoup4>=4.12.0
mysql-connector-python>=8.0.0
scrapy-user-agents>=0.1.0